feat(cab): Generate SMS hashes (#22)
### Description

For [ACT-1528](https://linear.app/valora/issue/ACT-1528/mvp-cab). See
[also](https://github.com/valora-inc/cloud-account-backup/pull/334).

### Test plan

<!-- Demonstrate the change is solid, or why it doesn't need testing.
Example: add any manual testing steps or scenarios (if not obvious),
screenshots / videos if the pull request changes the user interface.
-->

### Related issues

- Fixes [ACT-1528](https://linear.app/valora/issue/ACT-1528/mvp-cab).

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)