feat(statsig): Add Statsig API key (#25)
### Description

Part of
[ACT-1533](https://linear.app/valora/issue/ACT-1533/ff-statsig-and-segment).

<!-- A few sentences describing the overall effects and goals of the
pull request's commits.
What is the current behavior, and what is the updated/expected behavior
with this PR? -->

### Test plan

<!-- Demonstrate the change is solid, or why it doesn't need testing.
Example: add any manual testing steps or scenarios (if not obvious),
screenshots / videos if the pull request changes the user interface.
-->

### Related issues

- Fixes #[issue number here]

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)